+++ API_RESTFULL/backend/package.json
@@ -11,7 +11,7 @@
     "db:migrate-and-seed": "npx sequelize-cli db:migrate && npx sequelize-cli db:seed:all",
     "db:reset-and-start": "npm run db:reset && npm start",
     "db:migrate-and-start": "npx sequelize-cli db:migrate && npm start",
-    "start": "node ./index.js"
+    "start": "node ./ws-server.js"
   },
   "author": "mansour | cynthia | isaac",
   "license": "ISC",
